Refactor code
- Remove unnecessary schema
- Fixup in types: remove inheritance and improve naming
- Refactor utility code into accessor code/file
- Update usage in front-end
