--- backend/src/backend/primary/routers/well_completion/schemas.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-
-from src.services.utils.well_completion_utils import WellCompletionDataSet
-
-
-class WellCompletionData(BaseModel):
-    json_data: WellCompletionDataSet
